+++ meuporg-client/source/Gui/Hud.cpp
@@ -29,7 +29,7 @@ void Hud::init()
 
     healthText.setFont(fonts->get(ResourceId::MONOF_56));
     healthText.setCharacterSize(16);
-    healthText.setPosition(135, 35 - healthText.getGlobalBounds().height/2);
+    healthText.setPosition(135, (int)(35 - healthText.getGlobalBounds().height/2));
     healthText.setFillColor(sf::Color::White);
 
     xpBar.setPosition(sf::Vector2f(25, 80));
@@ -37,7 +37,7 @@ void Hud::init()
 
     xpText.setFont(fonts->get(ResourceId::MONOF_56));
     xpText.setCharacterSize(16);
-    xpText.setPosition(135, 105 - healthText.getGlobalBounds().height/2);
+    xpText.setPosition(135, (int)(105 - healthText.getGlobalBounds().height/2));
     xpText.setFillColor(sf::Color::White);
 
     background.setPosition(0, 0);
